Move ms fixture to conftest.py
fixtures in conftest.py are discoverable by pytest for use in all tests
